requirements.txt
asgiref==3.3.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.4
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
comtypes==1.1.7
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-cors-headers==3.6.0
django-import-export==2.5.0
djangorestframework==3.12.2
entrypoints==0.3
et-xmlfile==1.0.1
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupPy==1.14
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numpy==1.18.1
odfpy==1.4.1
opencv-python==4.1.2.30
openpyxl==3.0.6
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
pipenv==2020.8.13
playsound==1.2.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Pygments==2.6.1
PyMySQL==0.9.3
pyrsistent==0.16.0
python-dateutil==2.8.1
pyttsx3==2.81
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
sqlparse==0.4.1
tablib==3.0.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
virtualenv==16.7.9
virtualenv-clone==0.5.4
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.1.0
